requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
pillow
oemer
python-multipart
onnxruntime-gpu
nvidia-cublas-cu12
nvidia-cudnn-cu12
pyngrok
